Renamed linkDetector.js to patternDetector.js Switched conenction to TWITCHCLIENT 3 (disables join/part) Added ChannelUsersCache, which loads the api chat list and simulates join/part Added ChannelHostCache and associated events to track channel hosters Added linux scripts for running bot as a daemon Launching on linux creates a pid file Added launch-i.sh for linux to launch an interactive window without creating a pid file Redirected all output to custom classes com.gmt2001.Console.[in/out/err] (in/out passthrough to console) Output to com.gmt2001.Console.err is logged to stderr.txt Added custom DefaultUncaughtExceptionHandler, which logs to err and stacktrace.txt HttpRequest now supports headers HttpResponse now includes any exception thrown Changed timeout in TwitchAPIv3 to 2 seconds Added handler for invalid JSON response to TwitchAPIv3 TwitchAPIv3.UpdateChannel now prefers exact match on game name Added TwitchAPIv3.GetChatUsers and TwitchAPIv3.GetHostUsers Moved announcement globals into announcements.js Announcement minimum interval is now 1 minute and minimum messages can be disabled Added saving of next announcement vars across bot instances Removed ability for non-mods to pay points to create a command in addCommand Registered non-custom commands can not be overwritten by addCommand Command permission and counter are deleted when command is deleted Custom commands with a set permission now silently fail if the user does not have permission Added marathonCommand, providing a scheduling system intended for marathons Quote command displays number of quotes when an invalid quote number is specified Fixed randomCommand Moved rollCommand globals into rollCommand.js Admins are exempt from rollCommand usage limit Removed user-specific responses from rollCommand Put a better error message for streamCommand failures Updated list of valid commercial lengths for auto/manual commercials Added ability to make the bot output the current auto commercial settings Auto commercials can now be disabled by only setting the first parameter, to 0 Last commercial is saved across bot restarts4 Top 10 now silently fails if points system is disabled Added twitch and twitter commands with permissions on who can use it, including specified lists of users Added follow train identification and announcing to followHandler Added hostHandler.js for announcing hosts Added try/catch around script loading init.js and scripts in the util folder are loaded manually, all other scripts are now automatically located and loaded $.hook.call now supports the alwaysrun parameter, which causes all scripts to be triggered even if disabled Timers are now handled by a new powerful system in init.js setInterval and setTimeout are overridden to use the new timer system Event ircChannelMessage redirects messages from twitchnotify and jtv to ircPrivateMessage Fixed setconnectedmessage The reconnect command is now a Moderator level command The module command now supports the status subcommand to check the status of an individual script Removed some useless kappaTriggers Added toggle for subscriber detection mode, adding support for twitchnotify When subscriber detection mode is set to twitchnotify only, disables api detection Moved all pointSystem globals into poinstSystem.js Added truerandom toggle to raffles which uses random.org Raffle announcement now dynamically adds extra info, such as points cost or followers only Raffle prize can be changed while raffle is open Once a user wins a raffle, they are removed from entrants list in case of a repick Added raffle subcommand count to get the current number of entrants Added raidSystem to announce raiders and track number of times they have raided Chat moderation system now supports allowing Casters or Subscribers to post links Chat moderation system can now moderate message spamming, symbols, repeating characters, and graphemes Moderation commands are now issued once and bans do not purge first Chat moderation will now only trigger on a single offense per message, starting with the most severe punishment Removed the big spam for chatmod usage in place of a simple list of vars Command list now accepts script name manually, as well as a permission group Added $.commandExists and $.isCustomCommand to command list Command list will now only output commands usable by the permission group of the invoking user File system $.deleteFile command now supports immidiate deletion Added $.findFiles and $.isDirectory to file system Added automatic spacing to fix issues sending < character Added ability to disable chat responses from the bot. Makes it so only timeout/ban/unban/mods/clear get sent to chat Random number generator changed to java.security.SecureRandom Improved handling of strings in $.array.contains Added logging functions for 4 log files: chat, detected links, events (notifications/mod actions), and errors Logs can be enabled or disabled and will be kept for a specified number of days Replaced all usages of str.length and str.length with $.strlen(str), which detects the correct one to use Updated link detection to reduce aggressiveness in some instances Moderator list is now detected by polling /mods every 10 minutes in addition to MODE +O Added ability to set a users permission group without a bot response (goes to log only) Subscribers are now detected and added to a subscriber list using SPECIALUSER for 10 seconds after a message Youtube player writes current song to a text file HTTPServer supports using PUT to simulate a chat message. This is insecure (no authentication) and is therefore limited to localhost Make PhantomBot a singleton Added ability to set a custom port for HTTPServer, custom IRC server address, and custom message limit per 30 seconds Bot now enables message sending based on both MODE +O and /mods and will use /mods as a check when MODE -O is received Updated command line launch to allow setting any botlogin parameters desired Disabled full follower list lookup Cache's will disable themselves for a short period if too many consecutive api failures occur Channel now uses a separate queue for timeout/ban/unban/clear/mods commands which is always processed first and sent reguardless of moderator status ScriptApi timer scheduler now uses a thread pool of 8 instead of a single thread

+++ src/com/scripts/util/fileSystem.js
@@ -49,10 +49,15 @@ $.touchFile = function(path) {
     }
 }
 
-$.deleteFile = function(path) {
+$.deleteFile = function(path, now) {
     try {
         var f = new java.io.File(path);
-        f.deleteOnExit();
+        
+        if (now) {
+            f['delete']();
+        } else {
+            f.deleteOnExit();
+        }
     } catch (e) {
         println("Failed to delete '" + path + "': " + e)
     }
@@ -63,7 +68,44 @@ $.fileExists = function(path) {
         var f = new java.io.File(path);
         return f.exists();
     } catch (e) {
-        println("Failed to delete '" + path + "': " + e)
+        return false;
+    }
+    
+    return false;
+}
+
+$.findFiles = function(directory, pattern) {
+    try {
+        var f = new java.io.File(directory);
+        
+        var ret = new Array();
+        
+        if (!f.isDirectory()) {
+            throw "not a valid directory";
+        } else {
+            var files = f.list();
+            
+            for (var i = 0; i < files.length; i++) {
+                if (files[i].indexOf(pattern) != -1) {
+                    ret.push(files[i]);
+                }
+            }
+            
+            return ret;
+        }
+    } catch (e) {
+        println("Failed to search in '" + directory + "': " + e)
+    }
+    
+    return new Array();
+}
+
+$.isDirectory = function(path) {
+    try {
+        var f = new java.io.File(path);
+        return f.isDirectory();
+    } catch (e) {
+        return false;
     }
     
     return false;
